requirements.txt
# Database connectivity
pymysql==1.1.0

# Data handling (using latest version for Python 3.13 compatibility)
pandas>=2.2.0

# Environment variables (for secure credential management)
python-dotenv==1.0.0

scikit-learn==1.4.2
numpy==1.26.4
xlsxwriter==3.2.9
openpyxl==3.1.5
